Extract SqlConsole into separate class
