+++ hazelcast/src/main/java/com/hazelcast/client/console/ClientConsoleApp.java
@@ -1537,7 +1537,7 @@ public class ClientConsoleApp implements EntryListener, ItemListener, MessageLis
      * Starts the test application. It loads the client configuration using the resolution logic as described in
      * {@link HazelcastClient#newHazelcastClient()}.
      */
-    public static void main(String[] args) {
+    public static void run(String[] args) {
         HazelcastInstance client = HazelcastClient.newHazelcastClient();
         ClientConsoleApp clientConsoleApp = new ClientConsoleApp(client, System.out);
         clientConsoleApp.start(args);
